requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
groq>=0.4.0
numpy>=1.24.0